Refactor/utils
