+++ src/components/AppLink/AppLink.tsx
@@ -1,6 +1,6 @@
 import { type ComponentProps } from "react";
 import Link from "next/link";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 type AppLinkVariant = "clear" | "filled" | "clear-zommed";
 type BackgroundColor = "black" | "light" | "clear" | "gray" | "white";
