+++ src/components/Container/Container.tsx
@@ -1,5 +1,5 @@
 import type { HTMLAttributes } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 type BackgroundColor = "white" | "dark" | "light";
