+++ src/components/EmailSignUpForm/EmailSignUpForm.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 import Button from "../Button/Button";
 
 type VariantForm = "light" | "dark";
