+++ src/components/Icon/Icon.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 interface Props extends React.SVGProps<SVGSVGElement> {
   className?: string;
