+++ src/features/BurgerMenu/BurgerMenu.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import Button from "@/components/Button/Button";
 import CategoryLinks from "@/components/CategoryLinks/CategoryLinks";
 import Portal from "@/components/Portal/Portal";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 function BurgerMenu() {
   const [isOpen, setIsOpen] = useState(false);
